local repo update

+++ src/SDK/Globals/Globals.h
@@ -9,7 +9,7 @@ public:
 	int m_nScreenHeight = 0;
 	int m_nScreenWidht = 0;
 
-	int m_flCurAimFOV = 0;
+	float m_flCurAimFOV = 0.0f;
 
 	bool m_bIsInGame = false;
 	bool m_bIsGameUIVisible = false;
